Generalize.

+++ tests/ListTest.hs
@@ -3,7 +3,7 @@ import Prelude
 
 main :: IO ()
 main = do
-  putStrLn $ show $ sum [1,2,3]
-  putStrLn $ show $ product [1,2,3,4]
+  putStrLn $ show $ sum [1,2,3::Int]
+  putStrLn $ show $ product [1,2,3,4::Int]
   putStrLn $ show $ and [True]
   putStrLn $ show $ and [True, False]
